Model Running Functionality
Added runModel() and getModelData() functions. 

runModel does a 40-year simulation of pensPlan and returns the annual UAL, Assets, Liability, and UAL Growth Rates in a dictionary. 

getModelData runs multiple models, collects the dictionaries from each model and returns them as a list, alongside a dictionary of the mean annual values of each metric.

Both functions include code that will visualize the data in a graph that opens in a new tab, but is not saved anywhere at the moment.
Both functions also include parameters for setting the volatility, contribution rate, and employment growth of the model(s).

Also added variables for keeping track of the above metrics to pensPlan and pensPop.
Discount Rate is now also a mutable input rather than a static value.